Clean up project

+++ backend/src/models.py
@@ -1,10 +1,11 @@
-from sqlalchemy import Column, Integer, String, ARRAY
+from sqlalchemy import Column, Integer, String, Float, ARRAY
 from sqlalchemy.ext.declarative import declarative_base
 
-Base  = declarative_base()
+Base = declarative_base()
+
 
 class InstructionRecord(Base):
     __tablename__ = 'instruction_record'
-    id  = Column(Integer, primary_key=True, index=True)
+    id = Column(Integer, primary_key=True, index=True)
     instruction = Column(ARRAY(String), nullable=False, unique=True)
-    result = Column(Integer, nullable=True)
+    result = Column(Float, nullable=True)
